# docs/swagger.yaml
basePath: /
definitions:
  handlers.CheckEmailResponse:
    description: Email presence and verification status
    properties:
      exists:
        example: true
        type: boolean
      verified:
        example: true
        type: boolean
    type: object
...
  handlers.EngineEventPayload:
    properties:
      actor:
        description: 'One of: DEV, USER, HACKER'
        example: USER
        type: string
      event:
        description: Custom event identifier
        example: custom_event
        type: string
      lock:
        description: Whether to trigger lockdown
        example: false
        type: boolean
      message:
        description: Human-readable description
        example: Manual log injection
        type: string
    type: object
  handlers.GenericMessageResponse:
    description: Standard success message response
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.PasswordResetConfirm:
    properties:
      password:
        example: SuperSecurePassword123!
        type: string
      token:
        example: abc123token
        type: string
    type: object
  handlers.PasswordResetRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
...
  handlers.SetupPasswordRequest:
    properties:
      password:
        example: StrongPassword123!
        type: string
      token:
        example: verify-token-abc123
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      email_verified:
        description: <-- Added email verification info
        type: boolean
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  observer.Actor:
    enum:
    - DEV
    - USER
    - HACKER
    type: string
    x-enum-varnames:
    - DEV
    - USER
    - HACKER
  observer.ThreatLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-comments:
      LEVEL_LOCKDOWN: 4 - System-level tampering, triggers lockdown
      LEVEL_MEAN: 3 - Mid-tier severity, often for hacks
      LEVEL_MOCK: 2 - Snarky, non-critical
      LEVEL_NONE: "0"
      LEVEL_SOFT: 1 - Minor mistake, gentle pushback
    x-enum-varnames:
    - LEVEL_NONE
    - LEVEL_SOFT
    - LEVEL_MOCK
    - LEVEL_MEAN
    - LEVEL_LOCKDOWN
  observer.TracePayload:
    properties:
      actor:
        allOf:
        - $ref: '#/definitions/observer.Actor'
        description: Who triggered it
        example: HACKER
      event:
        description: Event type
        example: tamper
        type: string
      id:
        description: Unique trace ID
        example: abc123
        type: string
      lock:
        description: Lockdown trigger flag
        example: false
        type: boolean
      message:
        description: Event message
        example: unauthorized access detected
        type: string
      severity:
        allOf:
        - $ref: '#/definitions/observer.ThreatLevel'
        description: ThreatLevel (as int for frontend)
        example: 3
      timestamp:
        description: UTC time
        example: "2025-05-16T12:30:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Secure backend API built with Go, Keycloak, and SQLite.
  title: PeithoSecure Lite API
  version: "1.0"
paths:
...
  /api/v1/auth/check:
    get:
      description: Determines if the email exists in the system and whether it has
        been verified
      parameters:
      - description: Email address to check
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CheckEmailResponse'
        "400":
          description: Invalid email
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Check if email exists and is verified
      tags:
      - Auth
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Logs in the user and returns a Keycloak-issued JWT token pair along
        with email verification status
      parameters:
      - description: Username and password payload
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Malformed request or JSON parsing failed
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "429":
          description: Too many failed login attempts
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Authenticate user credentials
      tags:
      - auth
...
  /api/v1/auth/request-password-reset:
    post:
      consumes:
      - application/json
      description: Sends a password reset email to the user with a one-time token
      parameters:
      - description: Email to receive reset link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GenericMessageResponse'
        "400":
          description: Missing or invalid email
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Internal error or email send failed
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Request password reset link
      tags:
      - auth
...
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Accepts a one-time token and new password, and resets the account
        password
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GenericMessageResponse'
        "400":
          description: Invalid or missing token/password
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Internal error or reset failure
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Reset password using token
      tags:
      - auth
  /api/v1/auth/secure-sample:
    get:
      description: Returns a success message for authenticated users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/observer.TracePayload'
      security:
      - BearerAuth: []
      summary: Sample protected route
      tags:
      - Auth
...
  /api/v1/auth/setup-password:
    post:
      consumes:
      - application/json
      description: Sets initial password after email verification, creates Keycloak
        user, and writes PQC license
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetupPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GenericMessageResponse'
        "400":
          description: Missing or invalid input
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "409":
          description: Already verified or setup attempted twice
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Server, Keycloak, or license error
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Finalize user account setup
      tags:
      - auth
...
  /api/v1/events/log:
    post:
      consumes:
      - application/json
      description: Log a custom trace event (devtools, external agent, honeypot, etc.)
      parameters:
      - description: Custom trace event to log
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.EngineEventPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Event logged successfully
          schema:
            $ref: '#/definitions/handlers.GenericMessageResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "422":
          description: Unknown actor type
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Submit system trace event
      tags:
      - System
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"